+++ common/Node.hpp
@@ -2,18 +2,18 @@
 #define NODE_HPP_
 class Node {
 private:
-	double _lng;
-	double _lat;
+	long _lng;
+	long _lat;
 public:
 	Node(){}
-	Node(const double lng, const double lat) :
+	Node(const long lng, const long lat) :
 			_lng(lng), _lat(lat) {
 	}
 	static double lowerDistanceBound(const Node &node1, const Node &node2);
-	double getLat() const {
+	long getLat() const {
 		return _lat;
 	}
-	double getLng() const {
+	long getLng() const {
 		return _lng;
 	}
 };
